Added testing path creation for firefox profiles; reorganized tests
